+++ spec/Bibs/HoldingsSpec.php
@@ -11,53 +11,22 @@ use spec\Scriptotek\Alma\SpecHelper;
 
 class HoldingsSpec extends ObjectBehavior
 {
-    public $sample = '{
-          "holding": [
-            {
-              "library": {
-                "value": "65144020000121",
-                "desc": "BURNS"
-              },
-              "location": {
-                "value": "UNASSIGNED",
-                "desc": "UNASSIGNED location"
-              },
-              "link": "/almaws/v1/bibs/99100383900121/holdings/2221159990000121",
-              "holding_id": "2221159990000121"
-            },
-            {
-              "library": {
-                "value": "5867020000121",
-                "desc": "Main Library"
-              },
-              "location": {
-                "value": "MICR",
-                "desc": "Microforms"
-              },
-              "link": "/almaws/v1/bibs/99100383900121/holdings/2221410000000121",
-              "holding_id": "2221410000000121"
-            }
-          ],
-          "bib_data": {
-            "title": "Data base",
-            "issn": "0095-0033",
-            "publisher": "Association for Computing Machinery",
-            "link": "/almaws/v1/bibs/99100383900121",
-            "mms_id": 99100383900121,
-            "place_of_publication": "New York :",
-            "network_number": [
-              "(CONSER)  2011250895",
-              "(CKB)954926959913",
-              "(OCoLC)604911177"
-            ]
-          },
-          "total_record_count": 2
-        }';
-
     public function let(AlmaClient $client, Bib $bib)
     {
         $bib->mms_id = 'abc';
         $this->beConstructedWith($client, $bib);
+    }
+
+    public function it_is_initializable()
+    {
+        $this->shouldHaveType(Holdings::class);
+    }
+
+    protected function expectRequest($client)
+    {
+        $client->getJSON('/bibs/abc/holdings')
+            ->shouldBeCalled()
+            ->willReturn(SpecHelper::getDummyData('holdings_response.json'));
     }
 
     public function it_provides_a_lazy_interface_to_holding_objects(AlmaClient $client, Bib $bib)
@@ -86,18 +55,14 @@ class HoldingsSpec extends ObjectBehavior
 
     public function it_is_countable(AlmaClient $client)
     {
-        $client->getJSON('/bibs/abc/holdings')
-            ->shouldBeCalled()
-            ->willReturn(json_decode($this->sample));
+        $this->expectRequest($client);
 
         $this->shouldHaveCount(2);
     }
 
     public function it_provides_an_iterator_interface_to_holding_objects(AlmaClient $client)
     {
-        $client->getJSON('/bibs/abc/holdings')
-            ->shouldBeCalled()
-            ->willReturn(json_decode($this->sample));
+        $this->expectRequest($client);
 
         $this->shouldImplement('Iterator');
 
@@ -106,9 +71,7 @@ class HoldingsSpec extends ObjectBehavior
 
     public function it_provides_basic_data_without_loading_the_full_record(AlmaClient $client)
     {
-        $client->getJSON('/bibs/abc/holdings')
-            ->shouldBeCalledTimes(1)
-            ->willReturn(json_decode($this->sample));
+        $this->expectRequest($client);
 
         $this->shouldHaveCount(2);
         $this->all()[0]->shouldHaveType(Holding::class);