+++ src/Bibs/ElectronicCollections.php
@@ -1,0 +1,69 @@
+<?php
+
+namespace Scriptotek\Alma\Bibs;
+
+use Scriptotek\Alma\Client;
+use Scriptotek\Alma\Electronic\Collection;
+use Scriptotek\Alma\Model\IterableCollection;
+use Scriptotek\Alma\Model\LazyResourceList;
+use Scriptotek\Alma\Model\ReadOnlyArrayAccess;
+
+class ElectronicCollections extends LazyResourceList implements \Countable, \Iterator, \ArrayAccess
+{
+    use ReadOnlyArrayAccess;
+    use IterableCollection;
+
+    /**
+     * The Bib this ElectronicCollections list belongs to.
+     *
+     * @var Bib
+     */
+    public $bib;
+
+    /**
+     * ElectronicCollections constructor.
+     *
+     * @param Client $client
+     * @param Bib    $bib
+     */
+    public function __construct(Client $client, Bib $bib)
+    {
+        parent::__construct($client, 'electronic_collection');
+        $this->bib = $bib;
+    }
+
+    /**
+     * Get a single electronic collection record by id.
+     *
+     * @param string $id
+     *
+     * @return Representation
+     */
+    public function get($id)
+    {
+        return Collection::make($this->client, $id);
+    }
+
+    /**
+     * Convert a data element to a resource object.
+     *
+     * @param $data
+     *
+     * @return Representation
+     */
+    protected function convertToResource($data)
+    {
+        return Collection::make($this->client, $data->id)
+            ->init($data);
+    }
+
+    /**
+     * Generate the base URL for this resource.
+     *
+     * @return string
+     */
+    protected function urlBase()
+    {
+        return "/bibs/{$this->bib->mms_id}/e-collections";
+    }
+}